Added tags support

+++ bin/diary.sh
@@ -13,19 +13,43 @@ set -euo pipefail
 readonly POST_HEADER="---
 title: \"{{TITLE}}\"
 date: {{DATE}}
+tags: [{{TAGS}}]
 ---"
 
 generate_post() {
-    local content=$1
+    local file=$1
 
-    date=$(basename "$content" | sed 's/\.md$//')
-    title=$(grep -m 1 "^# " "$content" | sed 's/^# \+//')
-    text=$(sed '0,/^# /{/^# /d;}' "$content")
+    date=$(basename "$file" | sed 's/\.md$//')
+    title=$(grep -m 1 "^# " "$file" | sed 's/^# \+//')
+    text=$(sed '0,/^# /{/^# /d;}' "$file")
+    tags=$(get_tags "$file")
 
     echo "$POST_HEADER" \
         | sed "s|{{TITLE}}|$title|" \
-        | sed "s|{{DATE}}|$date|"
+        | sed "s|{{DATE}}|$date|" \
+        | sed "s|{{TAGS}}|$tags|"
     echo "$text"
+}
+
+get_tags() {
+    local file=$1
+    # Tag pattern.
+    readonly pattern=":[^ :]+:"
+    # How many lines to check from the beginning of file. Used to reduce false
+    # positives from the content such as URLs, etc.
+    readonly tags_offset=5
+
+    local header
+    header=$(head -n"$tags_offset" "$file")
+    if [ -f "$file" ]; then
+        if [[ $( echo "$header" | grep -cE "$pattern" || true) != "0" ]]; then
+            echo "$header" \
+                | grep -oE "$pattern" "$file" \
+                | sed "s/://g" \
+                | xargs \
+                | sed "s/ /,/g"
+        fi
+    fi
 }
 
 
@@ -34,8 +58,11 @@ generate_post() {
 readonly TBA_TAG=":tba:"
 
 get_content_files() {
-    for entry in "$source_path"*.md
+    local source_path=$1
+    for entry in "$source_path"/*.md
     do
+        [[ -e "$entry" ]] || break
+
         local filename
         filename=$(basename "$entry")
 
@@ -45,7 +72,8 @@ get_content_files() {
         fi
 
         # Skip TBA posts.
-        if [[ $(grep "$TBA_TAG" "$entry" -c) != 0 ]]; then
+        if [[ $(grep -i "$TBA_TAG" "$entry" -c) != 0 ]]; then
+            echo "Skipping. Found $TBA_TAG tag: $entry" >&2
             continue
         fi
 
@@ -57,16 +85,14 @@ main() {
     local source_path=$1
     local dest_path=$2
 
-    rm -rf "$dest_path/*"
+    rm -rf "${dest_path:?}/*"
     mkdir -p "$dest_path"
 
-    for path in $(get_content_files)
-    do
+    for path in $(get_content_files "$source_path"); do
         local filename
         filename=$(basename "$path")
         generate_post "$path" > "$dest_path/$filename"
     done
-
 }
 
 main "$@"
